pyproject.toml
[project]
name = "scrapscope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "tensorflow[and-cuda]>=2.19.0",
    "uvicorn>=0.34.3",
]
